The old way we handled logging didn't work well for multiple packages.
We have moved the logging init code into a new package utils which all
the other packages can import.  The current pattern we are using is 1
logger per package, initialised in the init() function.  Tests can then
disable the output of the logger by changing the level in suite
initialiser.