We don't want to run out of file descriptors
HTTP Responses must have `Close` called on the body, to ensure the file
descriptor is closed.

This was causing the bridge to freeze. When an event was received from
LRSC it would be forwarded to the IoTF which would attempt to register
the device before forwarding the message. When the file descriptors are
exhausted the call to registerDevice will block and no further action
will take place.